Align model-coords
Sub-buffers must have a minimum alignment, so make sure this is obeyed.
Adds some dead space at the end of each model. This means check for total #
of particles in unpack_chromo_coords has to be removed.

Note ALLOC_HOST_PTR for sub-buffer parent must be disabled on Intel
(Beignet #98490)